more working on shared helpers module to keep code DRY 2
