+++ serverless/requirements-dev.txt
@@ -1,9 +1,28 @@
--r requirements.txt
-black
-build
-flake8
-isort
-# mypy
-pytest
-pylint
-pre-commit
+astroid==3.3.9
+black==25.1.0
+build==1.2.2.post1
+cfgv==3.4.0
+click==8.1.8
+dill==0.3.9
+distlib==0.3.9
+filelock==3.18.0
+flake8==7.2.0
+identify==2.6.9
+iniconfig==2.1.0
+isort==6.0.1
+mccabe==0.7.0
+mypy-extensions==1.0.0
+nodeenv==1.9.1
+packaging==24.2
+pathspec==0.12.1
+platformdirs==4.3.7
+pluggy==1.5.0
+pre_commit==4.2.0
+pycodestyle==2.13.0
+pyflakes==3.3.2
+pylint==3.3.6
+pyproject_hooks==1.2.0
+pytest==8.3.5
+PyYAML==6.0.2
+tomlkit==0.13.2
+virtualenv==20.30.0
