+++ serverless/requirements.txt
@@ -1,1 +1,7 @@
-boto3
+boto3==1.37.31
+botocore==1.37.31
+jmespath==1.0.1
+python-dateutil==2.9.0.post0
+s3transfer==0.11.4
+six==1.17.0
+urllib3==2.3.0
